+++ package.json
@@ -14,16 +14,25 @@
     "start": "concurrently \"npm run start:backend\" \"npm run start:frontend\" \"mongod\" "
   },
   "dependencies": {
+    "@types/history": "^4.6.1",
+    "@types/react-router": "^4.0.15",
+    "@types/react-router-dom": "^4.0.8",
+    "@types/react-router-redux": "^5.0.8",
     "bcrypt-nodejs": "0.0.3",
-    "body-parser": "^1.18.1",
+    "body-parser": "^1.18.2",
     "concurrently": "^3.5.0",
-    "express": "^4.15.4",
-    "material-ui": "^0.19.2",
-    "mongoose": "^4.11.11",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "express": "^4.16.2",
+    "history": "^4.7.2",
+    "material-ui": "^0.19.4",
+    "mongoose": "^4.12.3",
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-redux": "^5.0.6",
-    "react-tap-event-plugin": "^2.0.1",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "^5.0.0-alpha.8",
+    "react-tap-event-plugin": "^3.0.2",
     "redux": "^3.7.2",
     "redux-thunk": "^2.2.0"
   },
@@ -31,33 +40,33 @@
     "@types/body-parser": "^1.16.5",
     "@types/express": "^4.0.37",
     "@types/html-webpack-plugin": "^2.28.0",
-    "@types/material-ui": "^0.18.1",
-    "@types/mongoose": "^4.7.21",
-    "@types/node": "^8.0.24",
-    "@types/react": "^16.0.2",
-    "@types/react-dom": "^15.5.3",
+    "@types/material-ui": "^0.18.3",
+    "@types/mongoose": "^4.7.24",
+    "@types/node": "^8.0.46",
+    "@types/react": "^16.0.15",
+    "@types/react-dom": "^16.0.2",
     "@types/react-hot-loader": "^3.0.3",
-    "@types/react-redux": "^5.0.6",
+    "@types/react-redux": "^5.0.10",
     "@types/react-tap-event-plugin": "0.0.30",
-    "@types/webpack": "^3.0.9",
-    "@types/webpack-env": "^1.13.0",
+    "@types/webpack": "^3.0.13",
+    "@types/webpack-env": "^1.13.2",
     "awesome-typescript-loader": "^3.2.3",
-    "connect-mongo": "^1.3.2",
+    "connect-mongo": "^2.0.0",
     "cors": "^2.8.4",
-    "cross-env": "^5.0.5",
+    "cross-env": "^5.1.0",
     "express-session": "^1.15.6",
     "favicons-webpack-plugin": "^0.0.7",
-    "file-loader": "^0.11.2",
+    "file-loader": "^1.1.5",
     "foreman": "^2.0.0",
     "html-webpack-plugin": "^2.30.1",
     "nodemon": "^1.12.1",
-    "react-hot-loader": "^3.0.0-beta.7",
+    "react-hot-loader": "^3.1.1",
     "rimraf": "^2.6.1",
-    "source-map-loader": "^0.2.1",
-    "ts-loader": "^2.3.4",
+    "source-map-loader": "^0.2.2",
+    "ts-loader": "^3.0.3",
     "ts-node": "^3.3.0",
-    "typescript": "^2.4.2",
-    "webpack": "^3.5.5",
-    "webpack-dev-server": "^2.7.1"
+    "typescript": "^2.5.3",
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.2"
   }
 }
