+++ src/frontend/elements/CollectionCard.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
 import Styles from '../style/elementStyles';
+// import PropTypes from 'prop-types';
 
 export default class CollectionCard extends React.Component<any, any> {
-    static propTypes = {
-        title: React.PropTypes.string.isRequired,
-    }
+    // static propTypes = {
+    //     title: PropTypes.string.isRequired,
+    // }
     public render() {
       return(
         <div className={Styles.collectionCardHead.class} style={Styles.collectionCardHead.style}>
