+++ src/frontend/elements/ResourceCard.tsx
@@ -1,11 +1,12 @@
 import * as React from 'react';
 import Styles from '../style/elementStyles';
+// import PropTypes from 'prop-types';
 
 export default class ResourceCard extends React.Component<any, any> {
-    static propTypes = {
-        title: React.PropTypes.string.isRequired,
-        subtitle: React.PropTypes.string.isRequired,
-    }
+    // static propTypes = {
+    //     title: PropTypes.string.isRequired,
+    //     subtitle: PropTypes.string.isRequired,
+    // }
     public render() {
       return(
         <div className={Styles.resourceCardHead.class} style={Styles.resourceCardHead.style}>
