+++ src/frontend/reducers/index.ts
@@ -5,6 +5,7 @@ import { user } from './user';
 import { users } from './users';
 import { contentUser } from './contentUser';
 import { addCollection } from './addCollection';
+import { routerReducer } from 'react-router-redux'
 
 // TODO: Add reducers
 
@@ -16,6 +17,7 @@ const rootReducer = combineReducers({
     users,  
     contentUser,
     addCollection, 
+    router: routerReducer
 });
   
 export default rootReducer
